+++ src/main/java/com/dominest/dominestbackend/domain/repeatschedule/repository/RepeatScheduleRepository.java
@@ -1,0 +1,7 @@
+package com.dominest.dominestbackend.domain.repeatschedule.repository;
+
+import com.dominest.dominestbackend.domain.repeatschedule.RepeatSchedule;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RepeatScheduleRepository extends JpaRepository<RepeatSchedule, Long> {
+}
